+++ mobile-app/lib/pages/run_admin/check_in_pack_page/check_in_pack_page.dart
@@ -360,7 +360,7 @@ class CheckInPackPage extends StatelessWidget {
                                             if (packMember.isPaid != 1) {
                                               unawaited(
                                                 AppScaffoldController.payForEvent(
-                                                  context,
+                                                  ScaffoldMessenger.of(context),
 
                                                   paymentBankTransfer.value,
                                                   index,
@@ -473,7 +473,7 @@ class CheckInPackPage extends StatelessWidget {
                                             if (packMember.isPaid != 1) {
                                               unawaited(
                                                 AppScaffoldController.payForEvent(
-                                                  context,
+                                                  ScaffoldMessenger.of(context),
 
                                                   paymentCash.value,
                                                   index,